+++ view/frontend/web/js/view/payment/method-renderer/cards.js
@@ -76,22 +76,6 @@ define(
                 return retVal;
             },
 
-            /**
-             * @returns {Boolean}
-             */
-            isVaultEnabled: function () {
-                return this.vaultEnabler.isVaultEnabled();
-            },
-
-            /**
-             * Returns vault code.
-             *
-             * @returns {String}
-             */
-            getVaultCode: function () {
-                return window.checkoutConfig.payment[this.getCode()].vault_code;
-            },
-
             getData: function () {
                 const data = this._super();