+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.1](https://github.com/unzerdev/magento2/compare/4.0.0..4.0.1)
+### Changed
+*  Invoice B2B component on checkout page
+### Added
+*  Support for Wero payment method
+*  Support for Unzer Direct Debit recurring payments
+
 ## [4.0.0](https://github.com/unzerdev/magento2/compare/3.2.9..4.0.0)
 ### Changed
 *  Migration from Unzer UI Component V1 to Unzer UI Component V2